gic: Move common GIC definitions to shared header
Share definitions between gic_pl390.h and gic_v3.h.
